+++ homematicip/device.py
@@ -6,7 +6,7 @@ from datetime import datetime
 from typing import Iterable
 
 from homematicip.base.enums import *
-from homematicip.base.helpers import get_functional_channel
+from homematicip.base.helpers import get_functional_channel, get_functional_channels
 from homematicip.base.HomeMaticIPObject import HomeMaticIPObject
 from homematicip.group import Group
 
@@ -40,6 +40,7 @@ class Device(HomeMaticIPObject):
         "IFeatureDeviceUndervoltage": ["deviceUndervoltage"],
         "IFeatureMulticastRouter": ["multicastRoutingEnabled"],
         "IFeatureDeviceIdentify": [],
+        "IFeatureProfilePeriodLimit": [],
     }
 
     def __init__(self, connection):
@@ -1409,6 +1410,34 @@ class WiredDimmer3(Dimmer):
     """HMIPW-DRD3 (Homematic IP Wired Dimming Actuator – 3x channels)"""
 
 
+class DinRailDimmer3(Dimmer):
+    """ HMIP-DRDI3 (Dimming Actuator Inbound 230V – 3x channels, 200W per channel) electrical DIN rail """
+
+    def __init__(self, connection):
+        super().__init__(connection)
+        self.c1dimLevel = 0.0
+        self.c2dimLevel = 0.0
+        self.c3dimLevel = 0.0
+
+    def from_json(self, js):
+        super().from_json(js)
+        channels = get_functional_channels("MULTI_MODE_INPUT_DIMMER_CHANNEL", js)
+        if channels:
+            self.c1dimLevel = channels[0]["dimLevel"]
+            self.dimLevel = self.c1dimLevel
+            self.c2dimLevel = channels[1]["dimLevel"]
+            self.c3dimLevel = channels[2]["dimLevel"]
+
+
+    def __str__(self):
+        return "{} c1DimLevel({}) c2DimLevel({}) c3DimLevel({})".format(
+            super().__str__(),
+            self.c1dimLevel,
+            self.c2dimLevel,
+            self.c3dimLevel
+        )
+
+
 class WeatherSensor(Device):
     """ HMIP-SWO-B """
 
@@ -2051,9 +2080,9 @@ class RainSensor(Device):
 
     def __init__(self, connection):
         super().__init__(connection)
-        #:float:
-        self.raining = False
         #:bool:
+        self.raining = False
+        #:float:
         self.rainSensorSensitivity = 0.0
 
     def from_json(self, js):
@@ -2062,3 +2091,24 @@ class RainSensor(Device):
         if c:
             self.set_attr_from_dict("rainSensorSensitivity", c)
             self.set_attr_from_dict("raining", c)
+
+
+class TemperaturDifferenceSensor2(Device):
+    """ HmIP-STE2-PCB (Temperature Difference Sensors - 2x sensors) """
+
+    def __init__(self, connection):
+        super().__init__(connection)
+        #:float:
+        self.temperatureExternalDelta = 0.0
+        #:float:
+        self.temperatureExternalOne = 0.0
+        #:float:
+        self.temperatureExternalTwo = 0.0
+
+    def from_json(self, js):
+        super().from_json(js)
+        c = get_functional_channel("TEMPERATURE_SENSOR_2_EXTERNAL_DELTA_CHANNEL", js)
+        if c:
+            self.set_attr_from_dict("temperatureExternalDelta", c)
+            self.set_attr_from_dict("temperatureExternalOne", c)
+            self.set_attr_from_dict("temperatureExternalTwo", c)