+++ CHANGELOG.md
@@ -4,7 +4,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [UNRELEASED]
+
+## [1.0.1] 2021-05-27
+
+### ADDED
+- Devices
+    - [HMIP-DRDI3] (Dimming Actuator Inbound 230V – 3x channels, 200W per channel)
+- API
+  - Groups
+    - [PR: 410] Add support for channel parsing in SECURITY_ZONES and SECURITY_AND_ALARM
+
+## [1.0.0] 2021-04-05
+
+### ADDED
+- Devices
+    - [HmIP-STE2-PCB] (Temperature Difference Sensors - 2x sensors) 
+	
+### FIXED
+- API
+  - [BUG: 387] Groups were missing in the functional channels of devices
+  - [BUG: 398] 'ACCESS_CONTROL' isn't a valid option for class 'FunctionalHomeType'
+  - [BUG: 391] There is no class for device 'PUSH_BUTTON_FLAT' yet
 
 ## [0.13.1] 2021-01-23
 
@@ -363,7 +384,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - API
 -- Support for the new HMIP Cloud Update
 
-[Unreleased]: https://github.com/coreGreenberet/homematicip-rest-api/compare/0.13.1...HEAD
+[UNRELEASED]: https://github.com/coreGreenberet/homematicip-rest-api/compare/1.0.1...master
+[1.0.1]: https://github.com/coreGreenberet/homematicip-rest-api/compare/1.0.0...1.0.1
+[1.0.0]: https://github.com/coreGreenberet/homematicip-rest-api/compare/0.13.1...1.0.0
 [0.13.1]: https://github.com/coreGreenberet/homematicip-rest-api/compare/0.13.0...0.13.1
 [0.13.0]: https://github.com/coreGreenberet/homematicip-rest-api/compare/0.12.1...0.13.0
 [0.12.1]: https://github.com/coreGreenberet/homematicip-rest-api/compare/0.12.0...0.12.1
@@ -405,6 +428,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [PR: 320]: https://github.com/coreGreenberet/homematicip-rest-api/pull/320
 [BUG: 325]: https://github.com/coreGreenberet/homematicip-rest-api/issues/325
 [BUG: 342]: https://github.com/coreGreenberet/homematicip-rest-api/issues/342
+[BUG: 387]: https://github.com/coreGreenberet/homematicip-rest-api/issues/387
+[BUG: 391]: https://github.com/coreGreenberet/homematicip-rest-api/issues/391
+[BUG: 398]: https://github.com/coreGreenberet/homematicip-rest-api/issues/398
+[PR: 410]: https://github.com/coreGreenberet/homematicip-rest-api/pull/410
 
 [HMIP-FALMOT-C12]: https://github.com/coreGreenberet/homematicip-rest-api/issues/281
 [HMIP-WHS2]: https://github.com/coreGreenberet/homematicip-rest-api/issues/280
@@ -420,4 +447,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [HMIP-SRD]: https://github.com/coreGreenberet/homematicip-rest-api/issues/375
 [HMIP-WRCC2]: https://github.com/coreGreenberet/homematicip-rest-api/issues/373
 [HMIP-DRSI1]: https://github.com/coreGreenberet/homematicip-rest-api/issues/373
+[HmIP-STE2-PCB]: https://github.com/coreGreenberet/homematicip-rest-api/issues/386
 [activate_absence_permanent]: https://github.com/coreGreenberet/homematicip-rest-api/issues/357
+[HMIP-DRDI3]: https://github.com/coreGreenberet/homematicip-rest-api/pull/405